docs: create LICENSE.md
I added the initial license agreement content for the Naked Time source
code license. The source code license allows for non-commercial use of
the source code for educational and research purposes. The license also
allows for personal or internal use of the source code for commercial
purposes provided that the individual or organization using the source
code has a valid and active license for the Naked Time software.

The purpose of this license is to allow me to develop Naked Time "in
the open" while still protecting my intellectual property rights and the
ability to earn revenue and income from the use of the Naked Time
product. The license will also allow licensed customers the ability to
use the source code for their own purposes, provided that they have a
valid and active license for the Naked Time software.