docs: create README.md
I wrote a short description of the Naked Time product. The description
provides information about the goal and vision for the product.

I created the press release for the first release of the Naked Time
product. The goal of the press release is to provide focus and define
the scope of the first release of the product.

I wrote a getting started section for working with the source code for
the Naked Time product. I created a software requirements document that
provides information about what software is required to be installed in
the development environment.