ci: validate commit messages
I configured Naked Time to use Node.js 22.14.0. I stored the version
number in the .node-version file.

I updated the software requirements document to add Fast Node Manager
as a required pre-requisite for all platforms.

I created the package.json manifest file for Naked Time. I added the
metadata for the Naked Time product. Naked Time is not implemented using
Node.js, but it does use tools written in JavaScript that require
Node.js to be installed. The package.json file is used to manage the
development dependencies for the Naked Time product.

I added NPM dependencies on commitlint and Husky. commitlint is used to
validate that commit messages are properly formatted and follow the
Conventional Commits specification. Husky is used to run commitlint
automatically when a commit is made.

I created setup.sh and Setup.bat to automate project setup tasks. I
implemented them to install Node.js (if necessary) and configure the
correct Node.js version for the product, and then to install the NPM
dependencies and configure Husky to execute Git hooks for the
repository.

I updated README.md with instructions on how to run the setup.sh or
Setup.bat scripts.

I created the commitlint.yaml GitHub Actions workflow that will run
commitlint to validate commit messages for pull requests and for new
commits that are merged into the main branch.

I added commitizen to as an NPM dependency to the product. I configured
commitizen to run automatically when creating a commit to help the
developer to write and format the commit message properly.