build: create the development container
I created the definition of the development container for the Naked Time
product. The development container provides a standardized Linux
development environment that will make it easier to build and debug
Naked Time's cloud native services and APIs.

I added requirements to install Docker Desktop, Visual Studio Code, and
the Remote Development Extension Pack for Visual Studio Code. These
software packages are required in order to run the development container
for Naked Time.

I added documentation to README.md to instruct developers on when and
how to use the development container for working with the Naked Time
source code. I captured images and screen shots to illustrate how to
start the development container in Visual Studio Code or Jetbrains IDEs.

I added support for using Git LFS to manage large and unversionable
files in the repository and store them externally in a blob storage.