docs: create README.md
I added a vision for Naked Time to README.md. I added instructions on
how to clone the repository and use the go script to prepare the new
repository clone for local development.

I created the software requirements document listing the software that
needs to be installed on the developer's machine.