chore: use commitlint to validate commit messages
I added commitlint and Husky as dependencies to the product using NPM.
commitlint is being used to validate that commit messages are properly
formatted following the Conventional Commits specification and the
configured rules for the Naked Time product. Husky is used to run
commitlint for the commit-msg Git hook and prevent commits to the local
repository if commit messages are not properly formatted.