ci: validate commit messages
I added commitlint and Husky as dependencies to the product using NPM.
commitlint is being used to validate that commit messages are properly
formatted following the Conventional Commits specification and the
configured rules for the Naked Time product. Husky is used to run
commitlint for the commit-msg Git hook and prevent commits to the local
repository if commit messages are not properly formatted.

I created the commitlint.yaml GitHub Actions workflow. This workflow
will run commitlint to validate that commits pushed to the GitHub
repository are properly formatted. This workflow will run and validate
all commits in a pull request branch, and the latest commit pushed to
the main branch.

I added commitizen as an NPM dependency. I configured the
prepare-commit-msg hook to run commitizen to help the developer to build
a properly formatted commit message.