feat: perform a pomodoro
I created the Naked Time CLI program and implemented the feature to
perform a pomodoro. The program will start a timer that runs for 25
minutes while showing the user how much time is remaining in the
pomodoro. When the pomodoro completes, an alarm sound will play on the
speaker and a desktop notification will be displayed.

I integrated GORM and SQLite into the program to store a record of the
pomodoros performed by the user. The pomodoro record is created at the
start of the pomodoro and is updated with the end time and completion
status when the pomodoro completes or is cancelled.

Closes: #8