style: add license agreement to source files
I updated all of the source code files to add the text of the copyright
statement and the text of the license agreement to the header area.
This is meant to inform readers of the ownership and license of the
source code if the source file should somehow be separated from the
repository or redistributed separately for some purpose.