docs: create LICENSE.md
I added the copyright statement, the text of the license agreement, and
a FAQ section with answers to common questions readers may have about
the license and what they can do with the source code for Naked Time.

I updated all of the source code files to add the text of the copyright
statement and the text of the license agreement to the header area.
This is meant to inform readers of the ownership and license of the
source code if the source file should somehow be separated from the
repository or redistributed separately for some purpose.

Closes: #1