chore: create the development container
I created the development container specification that is used to build
the development container for the IDE. I configured the development
container to install Git, Git LFS, GitHub CLI, and Go programming
language support. I created a custom feature for the development
container to install the libasound2-dev package that is needed for
building the beep framework on Linux.