chore: create notify-send and kdialog features
I created local features to install notify-send and kdialog in the
development container. These tools are used when trying to send desktop
notifications on Linux. I did not add them to the development container
because the development container lacks the ability to display desktop
notifications. However, keeping these features could be useful, so I am
adding them to the repository for now.