chore: create the development container
I created the development container specification that is used to build
the development container for the IDE. I configured the development
container to install Git, Git LFS, GitHub CLI, and Go programming
language support. I created a custom feature for the development
container to install the libasound2-dev package that is needed for
building the beep framework on Linux.

I created local features to install notify-send and kdialog in the
development container. These tools are used when trying to send desktop
notifications on Linux. I did not add them to the development container
because the development container lacks the ability to display desktop
notifications. However, keeping these features could be useful, so I am
adding them to the repository for now.

Closes: #4