ci: change latest tag
I revised the Docker workflow to label the latest release image as
`latest` and to label the build from the `main` branch as `dev`. This
will ensure that `latest` always returns a production version of Naked
Time and will allow early adopters to test the latest features by
using the `dev` tag.