ci: add execute permissions to Docker container
I updated Dockerfile to add execute permissions to the time program in
the Docker container. The execute permissions are not transferred when
copying the program into the container.

I created a .dockerignore file to exclude unnecessary files from being
sent to the Docker daemon during the build process. This helps to
reduce the build context size and improve build performance.