fix: enable Git LFS downloads
I updated the actions/checkout GitHub Action to v5 for the macOS,
Windows, and Docker container GitHub Actions workflows. I also enabled
the flag to download Git LFS files. I think this is the cause of the
error that I am seeing where the alert sound is not playing when the
pomodoro is complete because the audio file is not being downloaded from
LFS.